[CLI] Verify network with provider
