+++ cast/tests/data/accounts/user1_representation
@@ -1,1 +1,1 @@
-SingleOwnerAccount { provider: JsonRpcClient { transport: HttpTransport { client: Client { accepts: Accepts, proxies: [Proxy(System({}), None)], referer: true, default_headers: {"accept": "*/*"} }, url: Url { scheme: "http", cannot_be_a_base: false, username: "", password: None, host: Some(Ipv4(127.0.0.1)), port: Some(5050), path: "/rpc", query: None, fragment: None } } }, signer: LocalWallet { private_key: SigningKey { secret_scalar: FieldElement { inner: 0x0000000000000000000000000000000055ae34c86281fbd19292c7e3bfdfceb4 } } }, address: FieldElement { inner: 0x0112b61669a84940af30544786a6b615755b192173c4665a7d38fb3aa0541cac }, chain_id: FieldElement { inner: 0x0000000000000000000000000000000000000000000000534e5f474f45524c49 }, block_id: Tag(Latest) }
+SingleOwnerAccount { provider: JsonRpcClient { transport: HttpTransport { client: Client { accepts: Accepts, proxies: [Proxy(System({}), None)], referer: true, default_headers: {"accept": "*/*"} }, url: Url { scheme: "http", cannot_be_a_base: false, username: "", password: None, host: Some(Ipv4(127.0.0.1)), port: Some(5055), path: "/rpc", query: None, fragment: None } } }, signer: LocalWallet { private_key: SigningKey { secret_scalar: FieldElement { inner: 0x0000000000000000000000000000000055ae34c86281fbd19292c7e3bfdfceb4 } } }, address: FieldElement { inner: 0x0112b61669a84940af30544786a6b615755b192173c4665a7d38fb3aa0541cac }, chain_id: FieldElement { inner: 0x0000000000000000000000000000000000000000000000534e5f474f45524c49 }, block_id: Tag(Latest) }
