+++ cast/tests/e2e/main.rs
@@ -1,1 +1,2 @@
+mod helper_functions;
 pub(crate) mod helpers;
